+++ packages/storage-plus/src/prefix.rs
@@ -23,7 +23,7 @@ where
 {
     pub fn new(top_name: &[u8], sub_names: &[&[u8]]) -> Self {
         // FIXME: we can use a custom function here, probably make this cleaner
-        let storage_prefix = nested_namespaces_with_key(&[top_names], sub_names, b"");
+        let storage_prefix = nested_namespaces_with_key(&[top_name], sub_names, b"");
         Prefix {
             storage_prefix,
             data: PhantomData,
